engine: extract esm/tile_cache.js
